% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA Function.R
\name{anva}
\alias{anva}
\title{The purpose of this function will be to run an anova.}
\usage{
anva(predictor, response, dataframe)
}
\arguments{
\item{predictor}{is the independent variable}

\item{response}{is the dependent variable}

\item{dataframe}{is the dataset}
}
\value{
a summary of an analysis of variance
}
\description{
The purpose of this function will be to run an anova.
}
