% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{est_shrub_mass}
\alias{est_shrub_mass}
\title{calculate density of a shrub}
\usage{
est_shrub_mass(volume)
}
\arguments{
\item{:}{height of shrub (numeric, default 1)}
}
\value{
: density of shrub (numeric, default 1)
}
\description{
calculate density of a shrub
}
