% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{est_shrub_mass_dim}
\alias{est_shrub_mass_dim}
\title{calculate mass of shrub}
\usage{
est_shrub_mass_dim(length, width, height)
}
\arguments{
\item{:}{length of shrub (numeric)}
}
\value{
: mass of shrub
}
\description{
calculate mass of shrub
}
