% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlinearregression.R
\name{linplot}
\alias{linplot}
\title{The purpose of this function is to plot a graph based on a previously run linear regression using the package.
export}
\usage{
linplot(x.var, y.var, dataframe)
}
\arguments{
\item{x.var}{is the independent variable}

\item{y.var}{is the dependent variable}

\item{dataframe}{is the dataset you are analyzing}
}
\value{
a linear model of Hyalella sp. survival
}
\description{
The purpose of this function is to plot a graph based on a previously run linear regression using the package.
export
}
